CompeteAI/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aiohttp==3.9.5
    # via
    #   langchain
    #   langchain-community
aiosignal==1.3.1
    # via aiohttp
altair==5.3.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
anyio==3.7.1
    # via
    #   httpx
    #   openai
attrs==23.2.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
blinker==1.8.2
    # via streamlit
cachetools==5.3.3
    # via streamlit
certifi==2024.6.2
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via streamlit
dataclasses-json==0.6.7
    # via langchain-community
distro==1.9.0
    # via openai
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
    # via streamlit
greenlet==3.0.3
    # via
    #   playwright
    #   sqlalchemy
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via openai
idna==3.7
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jinja2==3.1.4
    # via
    #   altair
    #   pydeck
jsonpatch==1.33
    # via langchain-core
jsonpointer==2.4
    # via jsonpatch
jsonschema==4.22.0
    # via altair
jsonschema-specifications==2023.12.1
    # via jsonschema
langchain==0.2.1
    # via
    #   -r requirements.in
    #   langchain-community
langchain-community==0.2.4
    # via -r requirements.in
langchain-core==0.2.3
    # via
    #   -r requirements.in
    #   langchain
    #   langchain-community
    #   langchain-openai
    #   langchain-text-splitters
langchain-openai==0.1.8
    # via -r requirements.in
langchain-text-splitters==0.2.1
    # via langchain
langsmith==0.1.75
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
marshmallow==3.21.3
    # via dataclasses-json
mdurl==0.1.2
    # via markdown-it-py
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
ndjson==0.3.1
    # via wandbox-api
numpy==1.26.4
    # via
    #   altair
    #   langchain
    #   langchain-community
    #   pandas
    #   pyarrow
    #   pydeck
    #   streamlit
openai==1.33.0
    # via langchain-openai
orjson==3.10.3
    # via langsmith
packaging==23.2
    # via
    #   altair
    #   langchain-core
    #   marshmallow
    #   streamlit
pandas==2.2.2
    # via
    #   altair
    #   streamlit
pillow==10.3.0
    # via streamlit
playwright==1.44.0
    # via -r requirements.in
protobuf==4.25.3
    # via streamlit
pyarrow==16.1.0
    # via streamlit
pydantic==2.7.3
    # via
    #   langchain
    #   langchain-core
    #   langsmith
    #   openai
pydantic-core==2.18.4
    # via pydantic
pydeck==0.9.1
    # via streamlit
pyee==11.1.0
    # via playwright
pygments==2.18.0
    # via rich
python-dateutil==2.9.0.post0
    # via pandas
pytz==2024.1
    # via pandas
pyyaml==6.0.1
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
    #   wandbox-api
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.5.15
    # via tiktoken
requests==2.32.3
    # via
    #   langchain
    #   langchain-community
    #   langsmith
    #   streamlit
    #   tiktoken
    #   wandbox-api
rich==13.7.1
    # via streamlit
rpds-py==0.18.1
    # via
    #   jsonschema
    #   referencing
six==1.16.0
    # via python-dateutil
smmap==5.0.1
    # via gitdb
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
sqlalchemy==2.0.30
    # via
    #   langchain
    #   langchain-community
streamlit==1.35.0
    # via -r requirements.in
tenacity==8.3.0
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
    #   streamlit
tiktoken==0.7.0
    # via langchain-openai
toml==0.10.2
    # via
    #   streamlit
    #   wandbox-api
toolz==0.12.1
    # via altair
tornado==6.4.1
    # via streamlit
tqdm==4.66.4
    # via openai
typing-extensions==4.12.1
    # via
    #   openai
    #   pydantic
    #   pydantic-core
    #   pyee
    #   sqlalchemy
    #   streamlit
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
tzdata==2024.1
    # via pandas
urllib3==2.2.1
    # via requests
wandbox-api==0.10.0
    # via -r requirements.in
watchdog==4.0.1
    # via streamlit
yarl==1.9.4
    # via aiohttp
